+++ lib/label/label.h
@@ -77,7 +77,7 @@ struct label_ops {
 	int (*read) (struct labeller * l, struct device * dev,
 		     void *label_buf,
 		     struct label_read_data *ld, struct label ** label,
-		     uint32_t *failed_flags);
+		     uint64_t *failed_flags);
 
 	/*
 	 * Additional consistency checks for the paranoid.
